make authorized requests by adding req.headers.Authorization user's token

+++ src/components/WorkoutDetails.js
@@ -1,20 +1,26 @@
 import React from "react";
 import { useWorkoutContext } from "../hooks/useWorkoutsContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 //date fns
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
+
 const WorkoutDetails = ({ workout }) => {
   const { dispatch } = useWorkoutContext();
+  const { user } = useAuthContext();
 
   const handleDelete = async (e) => {
+    if (!user) {
+      return;
+    }
     const response = await fetch(
       "http://localhost:4000/api/workouts/" + workout._id,
       {
         method: "DELETE",
-        // body: JSON.stringify(workout),
-        // headers: {
-        //   "Content-Type": "application/json",
-        // },
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: `Bearer ${user.token}`,
+        },
       }
     );
